boston/keras/requirements.txt
tensorflow==1.10.*
numpy==1.14.5
google-cloud-storage
setuptools==39.1.0